Fix permission bugs in team API
